backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
asyncpg>=0.27.0
psycopg2-binary>=2.9.5
redis>=4.5.0
alembic>=1.10.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0
fastapi-utils>=0.2.1